refactor(server): move headless markdown helpers into package
